+++ test_datastructure.c
@@ -9,7 +9,7 @@ void sparse_set_new__succeeds(void) {
   assert(s->count == 0);
 
   for (size_t i = 0; i < s->capacity; i++) {
-    assert(s->sparse[i] == SIZE_MAX);
+    assert(s->sparse[i] == UINT32_MAX);
   }
 
   sparse_set_free(s);
@@ -45,7 +45,7 @@ void sparse_set_remove__succeeds_with_last_element(void) {
   int r = sparse_set_remove(s, e);
   assert(r == 0);
   assert(s->count == 0);
-  assert(s->sparse[entity_get_index(e)] == SIZE_MAX);
+  assert(s->sparse[entity_get_index(e)] == UINT32_MAX);
 
   sparse_set_free(s);
 }
@@ -63,7 +63,7 @@ void sparse_set_remove__swaps_last_dense_array_element(void) {
   int r = sparse_set_remove(s, e);
   assert(r == 0);
   assert(s->dense[to_delete_dense_pos] == last_entity);
-  assert(s->sparse[e_idx] == SIZE_MAX);
+  assert(s->sparse[e_idx] == UINT32_MAX);
 
   sparse_set_free(s);
 }
@@ -73,7 +73,7 @@ void sparse_set_remove__fails_if_element_does_not_exist(void) {
   entity e = entity_new(1, 4);
   int r = sparse_set_remove(s, e);
   assert(r == -1);
-  assert(s->sparse[entity_get_index(e)] == SIZE_MAX);
+  assert(s->sparse[entity_get_index(e)] == UINT32_MAX);
   assert(s->count == 0);
 
   sparse_set_free(s);
